organization similarity calculator with addition checks
